assignment_9/requirements.txt
streamlit
pyrebase4
firebase-admin
python-dotenv
